use diskutil to pick APFS Container
The APFS disk identifier in a MacOS VM may not be disk1; in our use case (shapehq/tartelet runners on MacStadium hosts) it ends up being disk4.  This line might be better at determining which disk to select in non-github-provided runners (or if GH changes up their vm structure).

This should be tested and could be made more robust.